player: don't update progress when page is not focused
Fixes an issue where on Firefox, airsonic-refix uses ~50% CPU when
playing a track. Even when not focused.

The issue lies somewhere in the progress bar being updated form ontimeupdate

Looking at a profile:

	RefreshDriver tick (70%) -> Update the rendering paint (55%) -> Paint (54%)

Chrome seems not have the issue, maybe it's internally throttling when
not focused.

For Firefox, just not updating the store when the document is hidden
dropped usage down to 8% from 50% for me.